+++ packages/mass-tools/src/index.js
@@ -30,4 +30,4 @@ export { generateMFs } from 'mf-generator';
 
 export { atomSorter } from 'atom-sorter';
 export { mfFromEA } from 'mf-from-ea';
-export { OctoChemDB, ActiveOrNaturalSummarizer } from 'octochemdb';
+export { OctoChemDB, ActiveOrNaturalSummarizer, createTaxonomyTree } from 'octochemdb';