+++ packages/octochemdb/src/index.js
@@ -1,3 +1,4 @@
 export * from './OctoChemDB';
 export * from './summarizer/ActiveOrNaturalSummarizer.js';
 export * from './taxonomy/taxonomyComparator.js';
+export * from './taxonomy/createTaxonomyTree.js';